+++ CHANGELOG.md
@@ -85,3 +85,4 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Fixed disable JIT optimization related logic.
 - Fixed constant field (literal) evaluation logic.
 - Fixed entry breakpoint logic, will not double break in case some source breakpoint is also set to first line of Main() method.
+- Fixed function breakpoint logic, will not double break in case some source breakpoint is also set to first line of method.